% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSigScore.R
\name{calcSigScore}
\alias{calcSigScore}
\title{Function to identify breast cancer molecular subtypes using the Subtype Clustering Model}
\usage{
calcSigScore(x, data, annot, doMapping = FALSE, mapping = NULL, size = 0,
  cutoff = NA, signed = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{Matrix containing the gene(s) in the gene list in rows and at least three columns: "probe", "EntrezGene.ID" and "coefficient" standing for the name of the probe, the NCBI Entrez Gene id and the coefficient giving the direction and the strength of the association of each gene in the gene list.}

\item{data}{Matrix of gene expressions with samples in rows and probes in columns, dimnames being properly defined.}

\item{annot}{Matrix of annotations with at least one column named "EntrezGene.ID" (for ssp, scm, AIMS, and claudinLow models) or "Gene.Symbol" (for the intClust model), dimnames being properly defined.}

\item{doMapping}{TRUE if the mapping through Entrez Gene ids must be performed (in case of ambiguities, the most variant probe is kept for each gene), FALSE otherwise.}

\item{mapping}{Matrix with columns "EntrezGene.ID" and "probe" used to force the mapping such that the probes are not selected based on their variance}

\item{size}{Integer specifying the number of probes to be considered in  signature computation. The probes will be sorted by absolute value of coefficients.}

\item{cutoff}{Only the probes with coefficient greater than cutoff will be considered in signature computation.}

\item{signed}{TRUE if only the sign of the coefficient must be considered in signature computation, FALSE otherwise.}

\item{verbose}{TRUE to print informative messages, FALSE otherwise.}
}
\value{
a list with 3 elements. "score", which has the risk predictions/scores for each patient, "mapping", which contains info regarding the number of genes in x that were mapped to the genes in data, and "probe", which has the IDs of the probes that were successfully mapped.
}
\description{
This function identifies the breast cancer molecular subtypes using a Subtype Clustering Model
}
\examples{

eset = loadMetaData(cancerType = "ovarian", survivalMetric = "overall")[1]
esetData = t(eset$E.MTAB.386@assayData$exprs)
colnames(esetData) = eset$E.MTAB.386@featureData@data$EntrezGene.ID
geneEntrezIds = eset$E.MTAB.386@featureData@data$EntrezGene.ID[1:5]
probes = eset$E.MTAB.386@featureData@data$probeset[1:5]
geneDirecs = rep(1, length(geneEntrezIds))
sig = cbind("probe"=probes, "EntrezGene.ID"=geneEntrezIds, "coefficient"= geneDirecs)
rownames(sig) = probes
scoreList = calcSigScore(x=sig, esetData, annot=NULL, doMapping=FALSE, signed=TRUE)

}
