% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateGenes.R
\name{correlateGenes}
\alias{correlateGenes}
\title{Function to determine the correlations between genes of patients}
\usage{
correlateGenes(geneEntrezIds, cancerType, survivalMetric, method = "pearson",
  dataList = NULL)
}
\arguments{
\item{geneEntrezIds}{A character vector containing the entrez IDs for the genes in the signature being tested}

\item{cancerType}{a string specifying what cancer type to use in the analysis. The options are "breast" and "ovarian"}

\item{survivalMetric}{a string specifying what type of survival data to use in the analysis. If the cancerType is set to "ovarian"
then enter "overall" for overall survival and "relapse" for relapse free survival. If the cancerType is set to "breast" then the options
are "relapse" for relapse free survival, "overallMeta" for overall survival on the metabric study data, "overallTcga" for
overall survival on the TCGA study data, and "overall" for overall survival on the remaining studies (No Metabric and TCGA data).}

\item{method}{The method to be used to compute the correlation coefficient. Options are pearson (default), spearman, and kendall.}

\item{dataList}{a list of esets to be used in the analysis. By default this variable is NULL and the other inputs will be used to obtain the esets for the analysis. However, it is useful to provide
a dataList if one is running this function multiple times so that the runtime will be reduced, or if there is a specific set of esets one would like to use in their analysis.}
}
\value{
a matrix containing the pairwise correlations between the supplied genes with the rows and columns named after the entrez gene ids
}
\description{
This function returns a matrix containing the pairwise correlations for gene signatures according to the correlation between the signature
scores in the patients that would be used for a given analysis (cancerType and survivalMetric determine the patients used)
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
corMatrix = correlateGenes(geneEntrezIds, cancerType = "ovarian", survivalMetric = "overall")
}
