% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateGeneSigs.R
\name{correlateGeneSigs}
\alias{correlateGeneSigs}
\title{Function to determine the correlation between genes in a signature}
\usage{
correlateGeneSigs(sigInfoList, method = "pearson", subtypeName = NULL,
  genesRequired = 0.75)
}
\arguments{
\item{sigInfoList}{The list returned from the function getGenesProgValue.}

\item{method}{The method to be used to compute the correlation coefficient. Options are pearson (default), spearman, and kendall.}

\item{subtypeName}{Default is NULL, in which case all the patients in the esets are used. If a string corresponding to one of the subtypes from the analysis is supplied (subtypes found in names(sigInfoList$patientSurvData[[1]])) then only patients that are the subtype "subtypeName" will be used in the analysis.}

\item{genesRequired}{a number less than or equal to 1 and greater than 0 specifying how many genes must be present in a dataset from the signature in order for the scores of the patients
from that dataset to be used when calculating the correlation between 2 signatures}
}
\value{
a matrix containing the pairwise correlations between the supplied gene signatures with the rows and columns named after the names of the geneSigList elements
}
\description{
This function returns a matrix containing the pairwise correlations for the genes in a signature. The correlations returned are meta estimates as
the correlation is computed between the signatures within each dataset and then a meta estimate is calculated by using the correlations of the signatures in each dataset.
}
\examples{

geneSigList = list(c("KLK14","RHOX8","ADAMTS20","IDO1"), c("ADAMTS20","IDO1", "FAP"))
names(geneSigList) = c("4 Gene Signature", "3 Gene Signature")
geneDirecList = list(c(1, 1, -1, -1), c(-1, -1, 1))
sigInfoList = getGenesProgValue(geneSigList, geneDirecList, cancerType = "ovarian", subtype = "verhaak", survivalMetric = "overall")
corMatrix = correlateGeneSigs(sigInfoList)
}
