% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSurvivalReport.R
\name{createSurvivalReport}
\alias{createSurvivalReport}
\title{Function that assesses the prognostic value of a gene signature and returns a pdf of the results}
\usage{
createSurvivalReport(geneSigList, geneDirecList, cancerType, subtype,
  survivalMetric, numGroups = 2, dataNames = NULL,
  soloGeneAnalysis = FALSE, removeMid = 0, censorTime = 10,
  addBenchmarks = FALSE)
}
\arguments{
\item{geneSigList}{A list of character vectors containing the ensemble IDs, entrez IDs, or gene symbols for the genes signatures/genes to conduct a survival analysis on. Entrez IDs are 
 recommended as if gene symbols or ensemble Ids cant be mapped to entrez Ids then the gene will be omitted from the analysis. Note that The names of the
elements in the list will correspond to the names of the signatures in the report, and default names will be provided if the list elements are not named.}

\item{geneDirecList}{A list of numeric vectors composed of +1 and - 1 indicating the direction of association for each vector of genes supplied in the geneSigList vectors.
+1 implies that the expression of that gene in a patient will be added to the patients score and -1 implies the expression will be subtracted from their score.
If one is looking for high scores to be associated with good survival, than genes that high expression is believed to lead to good prognosis should be given +1 and genes that high expression is believed
to lead to bad prognosis should be given a minus 1.}

\item{cancerType}{A string representing the type of cancer one would like to test the gene signature on. Options are currently "ovarian" and "breast"}

\item{subtype}{a string representing the subtyping scheme the patients will be classified under. For breast cancer, the options are "scmgene", "scmod1", "scmod2", "pam50", "ssp2006", "ssp2003", "intClust", "AIMS", or "claudinLow".
For ovarian cancer the options are "verhaak", "bentink", "tothill", "helland", and "konecny".}

\item{survivalMetric}{a string specifying what type of survival data to use in the analysis. If the cancerType is set to "ovarian"
then enter "overall" for overall survival and "relapse" for relapse free survival. If the cancerType is set to "breast" then the options
are "relapse" for relapse free survival, "overallMeta" for overall survival on the metabric study data, "overallTcga" for
overall survival on the TCGA study data, "overall" for overall survival on the remaining studies (No Metabric and TCGA data), and metastasis for distant metastasis free survival.
Note that for all cancer types the option "hierarchy" is available and this survivalMetric parameter checks for relapse free, then distant metastasis, and then overall survival data
in each dataset and keeps the first survival data it finds. Note that the ovarian datasets do not have distant metastasis free survival info and for the breast analyses metabric and tcga are removed.}

\item{numGroups}{an integer specifying the number of groups the patients will be split into when generating survival curves and calculating the log rank p value. The value is 2 by default and cannot exceed 5.}

\item{dataNames}{a character vector specifying the names of the datasets to include in the analysis. The names must match the names in the column "Data Name" from the dataframe returned from the
obtainDataInfo function. Use loadMetaData followed by obtainDataInfo with the cancerType and survivalMetric of interest to get the appropriate data names in the obtainDataInfo table for your analysis. By default
all datasets are included.}

\item{soloGeneAnalysis}{a boolean specifying whether each unique gene in geneSigList should have an survival analysis conducted on it in order to assess its
prognostic value independent of its gene signature. Results will show up in a section called Individual Gene Survival Analysis and high scores in the survival curves correspond to high expression as the direction is defaulted to 1
to allow for easy comparison amongst all the genes. Default value is FALSE}

\item{removeMid}{a number greater than 0 and less than .5 specifying the fraction of the patients with scores/risk predictions in the middle of the patients scores to be removed when generating survival results
The default is 0, all patients used in the analysis}

\item{censorTime}{a number specifying the point in time (years) at which the survival data must be censored. The default is 10 years}

\item{addBenchmarks}{a boolean specifying whether to add known signatures from literature to the analysis for comparison to the provided signatures. Default value is FALSE}
}
\value{
A pdf called MetaGxReport in R's working directory at the time the function was called
}
\description{
This function runs a survival analysis on breast or ovarian cancer patients for a gene signature using overall survival or relapse freee survival data
}
\examples{
geneSigList = list(c("KLK14","RHOX8","ADAMTS20","IDO1"), c("ADAMTS20","IDO1", "FAP"))
names(geneSigList) = c("4 Gene Signature", "3 Gene Signature")
geneDirecList = list(c(1, 1, -1, -1), c(-1, -1, 1))
createSurvivalReport(geneSigList, geneDirecList, cancerType = "breast", subtype = "scmod2", survivalMetric = "relapse", addBenchmarks = TRUE)

}
