% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determMissingGenes.R
\name{determMissingGenes}
\alias{determMissingGenes}
\title{Function to determine which genes are present and absent from the datasets}
\usage{
determMissingGenes(geneIds, dataList)
}
\arguments{
\item{geneIds}{A character vector containing the ensemble IDs, entrez IDs, or gene symbols for the genes in the signature being tested}

\item{dataList}{a list containing the datasets that the genes will be searched for in}
}
\value{
a data frame with information pertaining to whether the genes were found in the supplied datasets, note that if the ID could not be mapped
to an entrez ID the supplied ID will be found in both the "Symbols of Missing Genes" and "Entrez IDs of Missing Genes" section as its ID is unknown
}
\description{
This function checks each dataset to determine whether the genes supplied are present for the patients
}
\examples{
dataList = loadMetaData("ovarian", "overall")
geneIds = c("KLK14","RHOX8","ADAMTS20","IDO1")
genesMissFrame = determMissingGenes(geneIds, dataList)

}
