% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremePatientSubset.R
\name{extremePatientSubset}
\alias{extremePatientSubset}
\title{Function to remove risk predictions from a list of survival info frames}
\usage{
extremePatientSubset(survInfoList, removeMid)
}
\arguments{
\item{removeMid}{a number greater than 0 and less than 1 specifying the fraction of the patients with scores/risk predictions in the middle of the patients scores to be removed when generating survival results
The default is 0, all patients used in the analysis}

\item{survInfo}{One of the data frames in the list returned from the function getPatientSurvivalData. Alternatively, a data frame with 5 columns called eventTime, eventStatus, scoreVals, scoreVals, numGenesPresent, and dataName that correspond to
the survival event times, the survival event's status, the score/risk prediction of the patient, the number of genes used to calculate the score and the name of the dataset that the patient is from.}
}
\value{
a data frame with information about the datasets, including the D index and standard error of the D index for each dataset
}
\description{
This function separates the patients into their respective datasets and then computes the D index for each dataset using the risk predictions, event times, and event occurence indicators of the patients
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
geneDirecs = c(1, 1, -1, -1)
survInfoList = getPatientSurvivalData(geneEntrezIds, geneDirecs, cancerType = "ovarian", subtype = "verhaak", survivalMetric = "overall")
survInfoList = extremePatientSubset(survInfoList, 0.33)

}
