% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestProbes.R
\name{getBestProbes}
\alias{getBestProbes}
\title{Function to identify which probes in an esets expression data that should be kept amongst those with the same IDs}
\usage{
getBestProbes(eset)
}
\arguments{
\item{eset}{an eset object}
}
\value{
a vector containing the row indices of the probes in the eset@assayData$exprs expresson matrix that have the highest IQR for each probe ID
}
\description{
This function returns the rows indices of the probes in eset@assayData$exprs that should be kept amongst probes with the same IDs. Probes with the highest IQR amongst those with the same IDs
are kept.
}
\examples{

eset = loadMetaData("ovarian", "overall")[[1]]
bestProbeRows = getBestProbes(eset)
expressMat = eset@assayData$exprs[bestProbeRows, ]

}
