% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreastSubtypes.R
\name{getBreastSubtypes}
\alias{getBreastSubtypes}
\title{Function to identify breast cancer molecular subtypes using the Subtype Clustering Model}
\usage{
getBreastSubtypes(data, subtypeModel = c("scmgene", "scmod1", "scmod2",
  "pam50", "ssp2006", "ssp2003", "intclust", "aims", "claudinlow"), annot,
  doMapping = FALSE)
}
\arguments{
\item{data}{Matrix of gene expressions with samples in columns and probes in rows, dimnames being properly defined.}

\item{subtypeModel}{a string representing the desired subtyping classification model. Subtyping classification model, can be either "scmgene", "scmod1", "scmod2", "pam50", "ssp2006", "ssp2003", "intClust", "AIMS", or "claudinLow".}

\item{annot}{Matrix of annotations with at least one column named "EntrezGene.ID" (for ssp, scm, AIMS, and claudinLow models) or "Gene.Symbol" (for the intClust model), dimnames being properly defined.}

\item{doMapping}{TRUE if the mapping through Entrez Gene ids must be performed (in case of ambiguities, the most variant probe is kept for each gene), FALSE otherwise.}
}
\value{
a list with 3 elements. "subtype", which has Subtypes identified by the subtyping classification model. "subtype.proba", which has the probabilities that a patient belongs to each subtype estimated by the subtyping classification model. "subtype.crisp", which has Crisp classes identified by the subtyping classification model.
}
\description{
This function identifies the breast cancer molecular subtypes using a Subtype Clustering Model
}
\examples{
eset = loadMetaData("breast", "overall")[[1]]
expressionData = eset@assayData$exprs
subtypePredictions = getBreastSubtypes(data = expressionData, subtypeModel = "scmod2", annot = eset@featureData@data, doMapping = TRUE)
}
