% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDindexOfDatasets.R
\name{getDindexOfDatasets}
\alias{getDindexOfDatasets}
\title{Function to compute the D index for risk predictions (and their standard errors) on each dataset that was used in the survival analysis}
\usage{
getDindexOfDatasets(survInfo, minPatients = 10)
}
\arguments{
\item{survInfo}{One of the data frames in the list returned from the function getPatientSurvivalData. Alternatively, a data frame with 7 columns called eventTime, eventStatus, scoreVals, scoreVals, numGenesPresent, dataName and patient ID that correspond to
the survival event times, the survival event's status, the score/risk prediction of the patient, the number of genes used to calculate the score, teh ID of the patient in the dataset and the name of the dataset that the patient is from.}

\item{minPatients}{an integer specifying the minimum number of patients/samples required to calculate a D.index. Default is 10}
}
\value{
a data frame with information about the datasets, including the D index and standard error of the D index for each dataset
}
\description{
This function separates the patients into their respective datasets and then computes the D index for each dataset using the risk predictions, event times, and event occurence indicators of the patients
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
geneDirecs = c(1, 1, -1, -1)
survInfoList = getPatientSurvivalData(geneEntrezIds, geneDirecs, cancerType = "ovarian", subtype = "verhaak", survivalMetric = "overall")
dIndexData = getDindexOfDatasets(survInfoList$IMR)
dIndexData
}
