% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEsetsProbesSubtypesEvents.R
\name{getEsetsProbesSubtypesEvents}
\alias{getEsetsProbesSubtypesEvents}
\title{Function to obtain a list of expression sets (eSets), their best probes, their patients subtypes, and the event time and status of the patients}
\usage{
getEsetsProbesSubtypesEvents(cancerType, survivalMetric, subtype,
  dataNames = NULL, remMissSubtypeData = TRUE)
}
\arguments{
\item{cancerType}{A string representing the type of cancer one would like to test the gene signature on. Options are currently "ovarian" and "breast"}

\item{survivalMetric}{a string specifying the type of survival analysis that will be investigated in the reports. Either "overall" for overall survival or "relapse" for relapse free survival}

\item{subtype}{a string representing the subtyping scheme the patients will be classified under. For breast cancer, the options are "scmgene", "scmod1", "scmod2", "pam50", "ssp2006", "ssp2003", "intClust", "AIMS", or "claudinLow".
For ovarian cancer the options are "verhaak", "bentink", "tothill", "helland", and "konecny".}

\item{dataNames}{a character vector specifying the names of the datasets to include in the analysis. The names must match the names in the column "Data Name" from the dataframe returned from the
obtainDataInfo function. Use loadMetaData followed by obtainDataInfo with the cancerType and survivalMetric of interest to get the appropriate data names in the obtainDataInfo table for your analysis. By default
all datasets are included.}

\item{remMissSubtypeData}{A boolean specifying whether datasets that do not contain patients with all of the subtypes in the subtyping scheme specified should be removed. TRUE by default and TRUE is required in order for the 
results to be used as inputs to the getPatientSurvivalData function when running a survival analysis on a given signatur eor gene}
}
\value{
a data frame with information about the esets supplied
}
\description{
This function returns a list of expression sets (eSets), a list of each eSets best probes (see getBestProbes for the nature of the results), 
a list of their event times (see getSurvEvent), a list of their event statuses (for the given survival metric specified), and adds a subtype element to the eSets (see getPatientSubtypes).
}
\examples{

esetsProbesEvents = getEsetsProbesSubtypesEvent("ovarian", "overall", "verhaak")

}
