% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Function to create a data frame with information pertaining to the signature}
\usage{
getGeneInfo(geneIds, geneDirec = NULL)
}
\arguments{
\item{geneIds}{A character vector containing the ensemble IDs, entrez IDs, or gene symbols for the genes in the signature being tested}

\item{geneDirecs}{A numeric vector composed of +1 and - 1 indicating the direction of association for each gene supplied in the geneIds vector. 
This is an optional argument for those that would like to append the direction of association of the genes to the dataframe (default is NULL).}
}
\value{
a data frame with information about the genes supplied
}
\description{
This function maps the geneIds supplied to ensemble IDs, entrez IDs, and gene symbols and returns a data frame with this information as well as information about the genes, when available
}
\examples{

geneIds = c("KLK14","RHOX8","ADAMTS20","IDO1")
geneDirecs = c(1, 1, -1, -1)
geneInfoFrame = getGeneInfo(geneIds, geneDirecs)
geneInfoFrame

}
