% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaDindexAndLogP.R
\name{getMetaDindexAndLogP}
\alias{getMetaDindexAndLogP}
\title{Function to obtain information about various esets}
\usage{
getMetaDindexAndLogP(survInfoList, numGroups, normalizeEsetScores = TRUE,
  numDigits = 2)
}
\arguments{
\item{survInfoList}{The list returned from the function getPatientSurvivalData. Alternatively, a list of data frames with 5 columns called eventTime, eventStatus, scoreVals, scoreVals, numGenesPresent, and dataName that correspond to
the survival event times, the survival event's status, the score/risk prediction of the patient, the number of genes used to calculate the score and the name of the dataset that the patient is from. 
an error will be thrown if the esets in dataList do not contain survival information for the requested event. Enter "overall" for
overall survival and "relapse" for relapse free survival.}

\item{numGroups}{an integer specifying the number of groups the patients will be split into when generating survival curves and calculating the log rank p value. The value is 2 by default and cannot exceed 5.}

\item{normalizeEsetScores}{A boolean specifying whether the groups used to determine the log rank p value should be determined by 1. Splitting the scores/risk predictions of the patients from each dataset in the number of groups specified or
2. normalizing the scores of the patients in each dataset such that the 2.5 percentile and 97.5 percentile for the patients scores in each dataset have the same value, then creating one vector of patient score and splitting this
vector of patient scores into the number of groups specified (TRUE). Normalization/TRUE is the default.}

\item{numDigits}{an integer specifying the number of digits for the values in the dataframe, Default is 2}
}
\value{
a data frame with the D index, D index confidence interval, D index p value, and log rank p value for each survival frame in the survInfoList
}
\description{
This function returns a data frame which describes the number of events (deceased or tumor recurrence), median time to an event, the number of patients, the number of genes, and the platform of an eset
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
geneDirecs = c(1, 1, -1, -1)
survInfoList = getPatientSurvivalData(geneEntrezIds, geneDirecs, cancerType = "ovarian", subtype = "verhaak", survivalMetric = "overall")
resultsTable = getMetaDindexAndLogP(survInfoList, 2)
resultsTable
}
