% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOvarianSubtypes.R
\name{getOvarianSubtypes}
\alias{getOvarianSubtypes}
\title{Function to identify ovarian cancer molecular subtypes}
\usage{
getOvarianSubtypes(eset, subtypeModel, intersectThresh = 0.75)
}
\arguments{
\item{eset}{an eset object}

\item{subtypeModel}{a string representing the desired subtyping classification model. Subtyping classification model, can be either "verhaak", "bentink", "tothill", "helland", or "konecny".}

\item{intersectThresh}{the fraction of genes from the eset that must be present in each verhaak subtype gene set in order for the patients in the eset to be classified according to the verhaak subtypes. Default value is 0.75 and the variable is not relevant for the other subtyping schemes.}
}
\value{
a list with 3 elements. "subtype", which has the subtype of each patient in the eSet for the given subtyping classification model. The second element varies and is specific to the suptyping scheme, it generally provides information regarding how well each patient fit into each subtyping scheme. Annot.eset has information about the eset that was sent.
}
\description{
This function identifies the breast cancer molecular subtypes using a Subtype Clustering Model
}
\examples{

eset = loadMetaData("ovarian", "overall")[[1]]
subtypeList = getOvarianSubtypes(eset, "verhaak")
}
