% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatientSubtypes.R
\name{getPatientSubtypes}
\alias{getPatientSubtypes}
\title{Function to determine the molecular subtypes of patients in an eSet according to their gene expression}
\usage{
getPatientSubtypes(eset, cancerType, subtype, intersectThresh = 0.75)
}
\arguments{
\item{eset}{an eset object}

\item{cancerType}{a string representing the cancer type of the patients in the eset object. Must be one of "ovarian" or "breast".}

\item{subtype}{a string representing the subtyping scheme the patients will be classified under. For breast cancer, the options are "scmgene", "scmod1", "scmod2", "pam50", "ssp2006", "ssp2003", "intclust", "AIMS", or "claudinLow".
For ovarian cancer the options are "verhaak", "bentink", "tothill", "helland", and "konecny".}

\item{intersectThresh}{the fraction of genes from the eset that must be present in each verhaak subtype gene set in order for the patients in the eset to be classified according to the verhaak subtypes. Default value is 0.75 and the variable is not relevant for the other subtyping schemes.}
}
\value{
the original eset with a new subtype factor (eset$subtype) whose j'th element corresponds to the subtype for the patient in the j'th column of the eSets expression data (eset@assayData$exprs[, j])
}
\description{
This function determines the molecular subtypes of patients in an eSet according to their gene expression and the subtyping scheme specified by the user.
}
\examples{

eset = loadMetaData("ovarian", "overall")[[1]]
eset$subtype
eset = getPatientSubtypes(eset, cancerType = "ovarian", subtype = "verhaak")
eset$subtype[1:10]
}
