% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvEventData.R
\name{getSurvEventData}
\alias{getSurvEventData}
\title{Function to obtain events and event times for a given survival metric in each dataset}
\usage{
getSurvEventData(dataList, survivalMetric)
}
\arguments{
\item{dataList}{a list containing the datasets to be used in the analysis}

\item{survivalMetric}{a string specifying what type of survival data to use in the analysis. If the cancerType is set to "ovarian"
then enter "overall" for overall survival and "relapse" for relapse free survival. If the cancerType is set to "breast" then the options
are "relapse" for relapse free survival, "overallMeta" for overall survival on the metabric study data, "overallTcga" for
overall survival on the TCGA study data, and "overall" for overall survival on the remaining studies (No Metabric and TCGA data).}
}
\value{
a list with two lists. One called eventList that indicates whether the event occurred or not for the patients of each dataset and the
second list, called eventTimeList, indicates the time (in years) that the events in eventList were observed relative to the patients first checkup
}
\description{
This function returns the event times and event statuses for the patients in the datasets
}
\examples{

dataList = loadMetaData(cancerType = "breast", survivalMetric = "overall")
survEventList = getSurvEventData(dataList, survivalMetric = "overall")
}
