% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvInfo.R
\name{getSurvInfo}
\alias{getSurvInfo}
\title{Function to obtain survival info and risk predictions for an expression set (eSet object)}
\usage{
getSurvInfo(eset, geneEntrezIds, geneDirecs, survivalMetric,
  subtypeName = NULL, censorTime = 10, bestProbes = NULL)
}
\arguments{
\item{eset}{an eset object}

\item{geneEntrezIds}{A character vector containing the entrez IDs for the genes in the signature being tested}

\item{geneDirecs}{A numeric vector composed of +1 and - 1 indicating the direction of association for each gene supplied in the geneEntrezIds vector.
+1 implies that the expression of that gene in a patient will be added to the patients score and -1 implies the expression will be subtracted from their score.
If one is looking for high scores to be associated with good survival, than genes that high expression is believed to lead to good prognosis should be given +1 and genes that high expression is believed
to lead to bad prognosis should be given a minus 1.}

\item{survivalMetric}{a string specifying the type of survival analysis that will be investigated in the reports. Either "overall" for overall survival or "relapse" for relapse free survival}

\item{subtypeName}{Default is NULL, in which case all the patients in the eset are used. If a string corresponding to one of the subtypes in the esets eset$subtypes vector is supplied then only patients that are the subtype "subtypeName" will be used in the analysis.}

\item{censorTime}{an integer specifying the point in time (years) at which the survival data must be censored. The default is 10 years}

\item{bestProbes}{a list of vectors containing the row indices of the probes in the eset@assayData$exprs expresson matrix that should be used. If bestprobes is NULL than getBestprobes will be used to select the probes that have the highest IQR for each probe ID}
}
\value{
a list with 3 elements. "subtype", which has Subtypes identified by the subtyping classification model. "subtype.proba", which has the probabilities that a patient belongs to each subtype estimated by the subtyping classification model. "subtype.crisp", which has Crisp classes identified by the subtyping classification model.
}
\description{
This function returns a data frame with survival info and risk predictions for the patients in a given expression set object
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
geneDirecs = c(1, 1, -1, -1)
esetsAndProbes = getEsetsProbesSubtypes("ovarian", "overall", "verhaak")
eset = esetsAndProbes$esets$E.MTAB.386
survInfo = getSurvInfo(eset, geneEntrezIds, geneDirecs, survivalMetric = "overall")

firstSubtype = as.character(esetsAndProbes$esets$E.MTAB.386$subtypes[1])
survInfoSubtype = getSurvInfo(eset, geneEntrezIds, geneDirecs, survivalMetric = "overall", subtypeName = firstSubtype)
}
