% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaData.R
\name{loadMetaData}
\alias{loadMetaData}
\title{Function to load data required for the survival reports}
\usage{
loadMetaData(cancerType, survivalMetric, dataNames = NULL)
}
\arguments{
\item{cancerType}{a string specifying what cancer type to use in the analysis. The options are "breast" and "ovarian"}

\item{survivalMetric}{a string specifying what type of survival data to use in the analysis. If the cancerType is set to "ovarian"
then enter "overall" for overall survival and "relapse" for relapse free survival. If the cancerType is set to "breast" then the options
are "relapse" for relapse free survival, "overallMeta" for overall survival on the metabric study data, "overallTcga" for
overall survival on the TCGA study data, "overall" for overall survival on the remaining studies (No Metabric and TCGA data), and metastasis for distant metastasis free survival.
Note that for all cancer types the option "hierarchy" is available and this survivalMetric parameter checks for relapse free, then distant metastasis, and then overall survival data
in each dataset and keeps the first survival data it finds. Note that the ovarian datasets do not have distant metastasis free survival info and for the breast analyses metabric and tcga are removed.}

\item{dataNames}{a character vector specifying the names of the datasets to include in the analysis. The names must match the names in the column "Data Name" from the dataframe returned from the
obtainDataInfo function. Use loadMetaData followed by obtainDataInfo with the cancerType and survivalMetric of interest to get the appropriate data names in the obtainDataInfo table for your analysis. By default
all datasets are included.}
}
\value{
a list containing the datasets that will be used in the analysis
}
\description{
This function returns the data required to generate a survival report according to the user inputs that define the cancer type and survival type being considered
}
\examples{

dataList = loadMetaData(cancerType = "ovarian", survivalMetric = "overall")
}
