% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeForestPlot.R
\name{makeForestPlot}
\alias{makeForestPlot}
\title{Function to generate a forest plot in order to visually compare the results form the survival analysis in the individually expression sets}
\usage{
makeForestPlot(survInfo, minPatients = 10, reportSize = FALSE)
}
\arguments{
\item{survInfo}{One of the data frames in the list returned from the function getPatientSurvivalData. Alternatively, a data frame with 6 columns called eventTime, eventStatus, scoreVals, scoreVals, numGenesPresent, dataName and patient ID that correspond to
the survival event times, the survival event's status, the score/risk prediction of the patient, the number of genes used to calculate the score, the name of the dataset that the patient is from and the ID of the patient int that dataset.}

\item{minPatients}{an integer specifying the minimum number of patients/samples required to calculate a D.index. Default is 10}

\item{reportSize}{a boolean indicating whether the size of the plot should be adjusted such that it is the optimal size for the pdf reports generated by createSurvivalReport. By default FALSE, and it should not be set to TRUE as it will likely make seeing the forest plot in R difficult.}
}
\value{
a forest plot with D indices determined by using the signature to calculate risk predictions for the patients in multiple datasets
}
\description{
This function creates a forest plot for the resulting D indices from using the signature to calculate risk predictions for the patients in multiple datasets
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
geneDirecs = c(1, 1, -1, -1)
survInfoList = getPatientSurvivalData(geneEntrezIds, geneDirecs, cancerType = "ovarian", subtype = "verhaak", survivalMetric = "overall")
makeForestPlot(survInfoList$`All Patients`)
}
