% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRiskBoxplots.R
\name{makeRiskBoxplots}
\alias{makeRiskBoxplots}
\title{Function to create a boxplot showing the risk predictions for the patients in each subtype}
\usage{
makeRiskBoxplots(survInfoList)
}
\arguments{
\item{survInfoList}{The list returned from the function getPatientSurvivalData. Alternatively, a list of data frames with 5 columns called eventTime, eventStatus, scoreVals, scoreVals, numGenesPresent, and dataName that correspond to
the survival event times, the survival event's status, the score/risk prediction of the patient, the number of genes used to calculate the score and the name of the dataset that the patient is from. 
The names of the dataframes in the list will be used for the boxplot names}
}
\value{
a boxplot with a p value from the Kruskal-Wallis rank sum test between the scores/risk predictions in the data frames of survInfoList
}
\description{
This function returns a data frame which describes the number of events (deceased or tumor recurrence), median time to an event, the number of patients, the number of genes, and the platform of an eset
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
geneDirecs = c(1, 1, -1, -1)
survInfoList = getPatientSurvivalData(geneEntrezIds, geneDirecs, cancerType = "ovarian", subtype = "verhaak", survivalMetric = "overall")
makeRiskBoxplots(survInfoList)

}
