% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSurvivalPlot.R
\name{makeSurvivalPlot}
\alias{makeSurvivalPlot}
\title{Function to generate a Kaplan-Meier survival curve using risk predictons and survival data}
\usage{
makeSurvivalPlot(survInfo, numGroups, normalizeEsetScores = TRUE,
  titleStr = "Kaplan-Meier Survival Curve", xLoc = 2)
}
\arguments{
\item{survInfo}{One of the data frames in the list returned from the function getPatientSurvivalData. Alternatively, a data frame with 6 columns called eventTime, eventStatus, scoreVals, scoreVals, numGenesPresent, dataName and patient ID that correspond to
the survival event times, the survival event's status, the score/risk prediction of the patient, the number of genes used to calculate the score, the name of the dataset that the patient is from and the ID of the patient int that dataset.}

\item{numGroups}{an integer specifying the number of groups the patients will be split into when generating survival curves and calculating the log rank p value. The value is 2 by default and cannot exceed 5.}

\item{normalizeEsetScores}{A boolean specifying whether the groups used to determine the log rank p value should be determined by 1. Splitting the scores/risk predictions of the patients from each dataset in the number of groups specified or
2. normalizing the scores of the patients in each dataset such that the 2.5 percentile and 97.5 percentile for the patients scores in each dataset have the same value, then creating one vector of patient score and splitting this
vector of patient score sinto the number of groups specified (TRUE). Normalization/TRUE is the default.}

\item{titleStr}{a string to be used as the title of the plot. If not supplied a default title will be used.}

\item{xLoc}{the location of the D index and lo rank p value along the x-axis in the plot. By default 2}
}
\value{
a Kaplan-Meier survival curve with the corresponding D index and log rank p value in the plot
}
\description{
This function generates a Kaplan-Meier survival curve using the risk predictons and survival data stored in the survInfo data frame
}
\examples{

geneEntrezIds = c("43847", "434768", "80070", "3620")
geneDirecs = c(1, 1, -1, -1)
survInfoList = getPatientSurvivalData(geneEntrezIds, geneDirecs, cancerType = "ovarian", subtype = "verhaak", survivalMetric = "overall")
makeSurvivalPlot(survInfoList$`All Patients`, numGroups = 2)
}
