% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainCancerSigs.R
\name{obtainCancerSigs}
\alias{obtainCancerSigs}
\title{Function to obtain cancer gene signatures from literature}
\usage{
obtainCancerSigs(cancerType)
}
\arguments{
\item{cancerType}{A string specifying what cancer to obtain signatures for. Options are currently "ovarian" and "breast"}
}
\value{
A list containing 2 lists of vectors. List 1 corresponds to the gene signatures and list 2 corresponds to the direction of association
for each gene in the gene signatures
}
\description{
This function returns known cancer gene signatures
}
\examples{

cancerSigs = obtainCancerSigs("ovarian")
}
