% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainDataInfo.R
\name{obtainDataInfo}
\alias{obtainDataInfo}
\title{Function to obtain information about various esets}
\usage{
obtainDataInfo(dataList, survivalMetric)
}
\arguments{
\item{dataList}{a list containing the datasets one would like info about}

\item{survivalMetric}{a string specifying whether to use overall survival or relapse free survival as the event of interest. Note that
an error will be thrown if the esets in dataList do not contain survival information for the requested event. Enter "overall" for
overall survival and "relapse" for relapse free survival.}
}
\value{
a data frame with information about the esets supplied
}
\description{
This function returns a data frame which describes the number of events (deceased or tumor recurrence), median time to an event, the number of patients, the number of genes, the platform of an eset, and the publication date (NA if no internet connection)
}
\examples{

dataList = loadMetaData(cancerType = "ovarian", survivalMetric = "overall")
dataInfoTab = obtainDataInfo(dataList, survivalMetric = "overall")
head(dataInfoTab)
}
